+++ components/UserAvatar.vue
@@ -1,0 +1,15 @@
+<template>
+    <div v-if="user" class="flex align-items-center">
+        <Avatar :image="user.meta?.avatarUrl" alt="user avatar" shape="circle" />
+        <span class="ml-2">
+            {{ user.meta?.displayName }}
+        </span>
+    </div>
+    <Message v-else severity="error">You may need to <i>includeRelated:['user']</i> to view this.</Message>
+</template>
+
+<script setup>
+const props = defineProps({
+    user: Object,
+});
+</script>